Mod: Included also to detect if more that one color is turned on.
Now, I have modified the code to also include if more that one color is detected on the traffic light.
For example if "Green and Yellow" or "Red and Yellow" are detected.
N.B: The code is not yet on the field.
Next, I will test this code if it's fully functional on the Toycar and merge it in to the other branch if succeeded.