fix: Both actuator and test-flight-E files are fixed
When running the actuator on the raspberry pi, the motor starts to move a by a little power without receiving a signal from the main model which in this case is the test flight. Now, it have fixed. Also, while using the function from @EsraelDawit-a code, the colors were wrong due to typo. Meaning that,  the function returns the color name starting with capital letter, and that is also fixed.
N.B. This code have successfully detected objects and traffic lights. The only thing left is now to demonstrate it  on the ground. 🏎️🚗
Test-flight-E is a success! 🍾🥳